Optimized memory usage by using a union, and other fixes
We use a union now for the 3 container types we store (std::string,
UniValue::Object and UniValue::Array).  We call the constructor
for the appropriate container explicitly when assigning to the UniValue or
giving the UniValue a type of VSTR, VNUM, VOBJ, or VARR.

If the union has an active member, its d'tor will be called on UniValue
destruction or if the UniValue underlying VType changes.

As per C++ spec, only 1 of the union members is active at once and we
were careful to ensure this is the case.

The switch to using a union makes each UniValue instance take 32 bytes
on 64-bit (previously it was 80 bytes). This should speed up processing
for larger data sets by preserving cache locality due to smaller memory
footprint, as well as just eating less memory when working with large JSON
documents (which is always a good idea).

Note that this deviates somewhat from the BCHN implementation now, which
doesn't do this.

Also in this commit:

- Renamed UniValue::NullUniValue -> UniValue::Null